writing unit and mvc test's according template branch NO-TRELLO-refactoring-session, did GlobalExceptionHandler according to template, and refactoring code.
