+++ core-service/src/main/java/com/crm/food/establishment/core/CoreServiceApplication.java
@@ -1,0 +1,14 @@
+package com.crm.food.establishment.core;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class CoreServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CoreServiceApplication.class, args);
+	}
+}
