+++ core-service/src/main/java/com/crm/food/establishment/core/ErrorTemplate.java
@@ -1,0 +1,7 @@
+package com.crm.food.establishment.core;
+public record ErrorTemplate(
+        String title,
+        String error
+        ) {
+
+}
