+++ core-service/src/main/java/com/crm/food/establishment/core/entity/Season.java
@@ -1,0 +1,8 @@
+package com.crm.food.establishment.core.entity;
+
+public enum Season {
+    SUMMER,
+    AUTUMN,
+    WINTER,
+    SPRING
+}
