+++ core-service/src/main/java/com/crm/food/establishment/core/service/MenuNotFoundException.java
@@ -1,0 +1,7 @@
+package com.crm.food.establishment.core.service;
+
+public class MenuNotFoundException extends RuntimeException {
+    public MenuNotFoundException(String message) {
+        super(message);
+    }
+}
