--- core-service/src/main/java/com/crmfoodestablishment/coreservice/CoreServiceApplication.java
@@ -1,14 +1,0 @@
-package com.crmfoodestablishment.coreservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-public class CoreServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CoreServiceApplication.class, args);
-	}
-}
