--- core-service/src/main/java/com/crmfoodestablishment/coreservice/entity/DishOrderId.java
@@ -1,21 +1,0 @@
-package com.crmfoodestablishment.coreservice.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-import java.io.Serializable;
-
-@Embeddable
-@EqualsAndHashCode
-@AllArgsConstructor
-@NoArgsConstructor
-public class DishOrderId implements Serializable {
-
-    @Column(name = "order_id", nullable = false)
-    private Long orderId;
-
-    @Column(name = "dish_id", nullable = false)
-    private Long dishId;
-}
