--- core-service/src/main/java/com/crmfoodestablishment/coreservice/entity/Season.java
@@ -1,8 +1,0 @@
-package com.crmfoodestablishment.coreservice.entity;
-
-public enum Season {
-    SUMMER,
-    AUTUMN,
-    WINTER,
-    SPRING
-}
