--- core-service/src/main/java/com/crmfoodestablishment/coreservice/service/MenuNotFoundException.java
@@ -1,7 +1,0 @@
-package com.crmfoodestablishment.coreservice.service;
-
-public class MenuNotFoundException extends RuntimeException {
-    public MenuNotFoundException(String message) {
-        super(message);
-    }
-}
