+++ core-service/src/main/resources/db/changelog/2023.08_initial_scripts.xml
@@ -32,6 +32,9 @@
             <column name="id" type="int" autoIncrement="true">
                 <constraints primaryKey="true"/>
             </column>
+            <column name="uuid" type="uuid">
+                <constraints unique="true" nullable="false"/>
+            </column>
             <column name="name" type="varchar(64)">
                 <constraints nullable="false" unique="true"/>
             </column>
