+++ core-service/src/test/java/com/crm/food/establishment/core/CoreServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.crm.food.establishment.core;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CoreServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
