--- core-service/src/test/java/com/crmfoodestablishment/coreservice/CoreServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.crmfoodestablishment.coreservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CoreServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
